Move view holders into there own package
