+++ src/main/java/org/quantumbadger/redreader/viewholders/VH.java
@@ -1,0 +1,11 @@
+package org.quantumbadger.redreader.viewholders;
+
+import android.support.v7.widget.RecyclerView;
+import android.view.View;
+
+public class VH extends RecyclerView.ViewHolder {
+
+	public VH(View itemView) {
+		super(itemView);
+	}
+}
